+++ models/mergeAmendments/Merge.php
@@ -61,7 +61,7 @@ class Merge
     private function mergeSimpleTextSection(MotionSection $section, MotionSection $origSection, Draft $draft): void
     {
         $paragraphs = [];
-        foreach ($origSection->getTextParagraphLines() as $paraNo => $para) {
+        foreach ($origSection->getTextParagraphLines(true) as $paraNo => $para) {
             $consolidated = $draft->paragraphs[$section->sectionId . '_' . $paraNo]->text;
             $consolidated = str_replace('<li>&nbsp;</li>', '', $consolidated);
             $paragraphs[] = $consolidated;