+++ History.md
@@ -18,6 +18,7 @@
 - It is now possible to deactivate the breadcrumb links in the appearance settings.
 - Agenda items, including their motions, can be hidden from the proposed procedure.
 - When creating a resolution using the "merge amendments"-functionality, there is now a "remove text"-checkbox above the motion reason to exclude the reason from the final resolution more easily.
+- Bugfix: if a motion section was empty (e.g. a motion that did not have a reason), that section could not be edited when merging amendments.
 - Bugfix: too long motion titles are now breaking the layout and some functionality less than before.
 
 ## Version 4.4.0 (2019-12-26)